fix(census): Moving selenium test to a new file ##283
Selenium tests have been moved to a new file test_selenium.py
