+++ decide/census/tests.py
@@ -4,7 +4,6 @@ from django.test import Client, TestCase
 from django.conf import settings
 from django.urls import reverse
 from rest_framework.test import APIClient
-from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from openpyxl import load_workbook
 from io import BytesIO
 from openpyxl import Workbook
@@ -21,7 +20,6 @@ from voting.models import Voting, Question, QuestionOption
 from base.models import Auth
 from base import mods
 from base.tests import BaseTestCase
-from datetime import datetime
 
 
 class CensusTestCase(BaseTestCase):
@@ -481,110 +479,6 @@ class CensusTestCase(BaseTestCase):
 
         self.assertEqual(deletion[0], 1)  # Comprobar que se eliminó un objeto
         self.assertEqual(0, Census.objects.count())  # Comprobar que no hay objetos Census
-
-class CensusTest(StaticLiveServerTestCase):
-    def setUp(self):
-        # Load base test functionality for decide
-        self.base = BaseTestCase()
-        self.base.setUp()
-
-        options = webdriver.ChromeOptions()
-        options.headless = True
-        self.driver = webdriver.Chrome(options=options)
-
-        super().setUp()
-
-    def tearDown(self):
-        super().tearDown()
-        self.driver.quit()
-
-        self.base.tearDown()
-
-    def createCensusSuccess(self):
-        self.cleaner.get(self.live_server_url + "/admin/login/?next=/admin/")
-        self.cleaner.set_window_size(1280, 720)
-
-        self.cleaner.find_element(By.ID, "id_username").click()
-        self.cleaner.find_element(By.ID, "id_username").send_keys("decide")
-
-        self.cleaner.find_element(By.ID, "id_password").click()
-        self.cleaner.find_element(By.ID, "id_password").send_keys("decide")
-
-        self.cleaner.find_element(By.ID, "id_password").send_keys("Keys.ENTER")
-
-        self.cleaner.get(self.live_server_url + "/admin/census/census/add")
-        now = datetime.now()
-        self.cleaner.find_element(By.ID, "id_voting_id").click()
-        self.cleaner.find_element(By.ID, "id_voting_id").send_keys(
-            now.strftime("%m%d%M%S")
-        )
-        self.cleaner.find_element(By.ID, "id_voter_id").click()
-        self.cleaner.find_element(By.ID, "id_voter_id").send_keys(
-            now.strftime("%m%d%M%S")
-        )
-        self.cleaner.find_element(By.NAME, "_save").click()
-
-        self.assertTrue(
-            self.cleaner.current_url == self.live_server_url + "/admin/census/census"
-        )
-
-    def createCensusEmptyError(self):
-        self.cleaner.get(self.live_server_url + "/admin/login/?next=/admin/")
-        self.cleaner.set_window_size(1280, 720)
-
-        self.cleaner.find_element(By.ID, "id_username").click()
-        self.cleaner.find_element(By.ID, "id_username").send_keys("decide")
-
-        self.cleaner.find_element(By.ID, "id_password").click()
-        self.cleaner.find_element(By.ID, "id_password").send_keys("decide")
-
-        self.cleaner.find_element(By.ID, "id_password").send_keys("Keys.ENTER")
-
-        self.cleaner.get(self.live_server_url + "/admin/census/census/add")
-
-        self.cleaner.find_element(By.NAME, "_save").click()
-
-        self.assertTrue(
-            self.cleaner.find_element_by_xpath(
-                "/html/body/div/div[3]/div/div[1]/div/form/div/p"
-            ).text
-            == "Please correct the errors below."
-        )
-        self.assertTrue(
-            self.cleaner.current_url
-            == self.live_server_url + "/admin/census/census/add"
-        )
-
-    def createCensusValueError(self):
-        self.cleaner.get(self.live_server_url + "/admin/login/?next=/admin/")
-        self.cleaner.set_window_size(1280, 720)
-
-        self.cleaner.find_element(By.ID, "id_username").click()
-        self.cleaner.find_element(By.ID, "id_username").send_keys("decide")
-
-        self.cleaner.find_element(By.ID, "id_password").click()
-        self.cleaner.find_element(By.ID, "id_password").send_keys("decide")
-
-        self.cleaner.find_element(By.ID, "id_password").send_keys("Keys.ENTER")
-
-        self.cleaner.get(self.live_server_url + "/admin/census/census/add")
-        now = datetime.now()
-        self.cleaner.find_element(By.ID, "id_voting_id").click()
-        self.cleaner.find_element(By.ID, "id_voting_id").send_keys("64654654654654")
-        self.cleaner.find_element(By.ID, "id_voter_id").click()
-        self.cleaner.find_element(By.ID, "id_voter_id").send_keys("64654654654654")
-        self.cleaner.find_element(By.NAME, "_save").click()
-
-        self.assertTrue(
-            self.cleaner.find_element_by_xpath(
-                "/html/body/div/div[3]/div/div[1]/div/form/div/p"
-            ).text
-            == "Please correct the errors below."
-        )
-        self.assertTrue(
-            self.cleaner.current_url
-            == self.live_server_url + "/admin/census/census/add"
-        )
 
 
 class CensusExportViewTest(BaseTestCase):
